🚨 Extract `project` from `models_config`
TG-2031
