+++ src/youwol/app/environment/models/__init__.py
@@ -7,5 +7,6 @@ from .defaults import *
 from .flow_switches import *
 from .models import *
 from .models_config import *
+from .models_project import *
 from .models_token_storage import *
 from .predefined_configs import *
